backend/requirements.txt
fastapi
uvicorn[standard]
httpx
openai
chromadb
pydantic
python-dotenv
apscheduler
python-multipart